make world file, lives, and prevent frog collision for finished frogs
